Added function to find and output roots of quadratic equation by coefficients
Added function that will get an array of coefficients of a quadratic
equation, check if the leading coefficient (a) is not equal to zero,
and solve quadratic equations according to the algorithm. After solving,
prints the details and the result to the console. The check of the
coefficient is made in order to avoid in the future a duplicate check
for different modes of the program.

Details on quadratic equations and an algorithm for solving:
https://yukhym.com/uk/matematika/kvadratne-rivnyannya.html