Created a function to get parameters from the console
Created a function getCoefFromConsole, that receives the parameter value
entered by the user from the console, and check if it matches the
expected format (real number). Also  created function
getParametersFromConsole, that receives 3 parameters (which we need
to solve the quadratic equation in the general case) using
getCoefFromConsole function, and returns them as a array.
The implementation of two different functions is made for decomposition,
and for a better understanding of the code.
A dependency readline-sync was used to read input from the console,
specifically this was chosen for its simplicity, which suits
our program. Also added a .gitignor file to hide node-modules folder.

readline-sync: https://www.npmjs.com/package/readline-sync