Added non-interactive mode of the program
Added index.js file, which implements the solution of a quadratic
equation, in case the correct path to the file with data in the
required format was specified in the console
(for example, 1.55/s3.22/s-4/n). The format check performe using a
regular expression. To read the file, the readFileSync module was used,
which corresponds to the synchronous way of writing the code of this
program. For test, a test.txt file was created, in which the
coefficients are specified in the correct format. Imported a module
with the solution of quadratic equations, for use in the index js,
also corrected the case with the wrong coefficient a
(now the program ends with code 1).